main menu, option menu + volume setting

+++ Assets/Scripts/ResourceData.cs
@@ -9,14 +9,14 @@ public class ResourceData : ScriptableObject
     [SerializeField] private string _name;
     [SerializeField] private float _health;
     [SerializeField] private Sprite _icon;
-    [SerializeField] private int price;
+    [SerializeField] private ulong price;
     #endregion
 
     #region Interface
     public string Name { get => _name; }
     public float Health { get => _health; }
     public Sprite Icon { get => _icon;  }
-    public int Price { get => price; }
+    public ulong Price { get => price; }
     #endregion
 }
